🧹 Refactor DoLCharacterSprite and add DoL feature parity for hallucination and parasites
